Finish Game without sound
Finished lvl4 by arranging the room a bit and adding turrets shooting
bullets. Also arranged the mainmenu but changing the picture and
changing the camera to a more 2D one. Start button was assigned to 'S'
and quit was added and assigned to 'Q'. Level Number added to all of the
scenes at the top of the screen.